EMBCESSMOD-720: Restriction Page UI - Registrant Portal - Non-verified Registration and Evacuation
