--- evacuees/src/UI/embc-evacuee/src/app/restriction/restriction.component.html
@@ -1,25 +1,0 @@
-
-    <div class="row heading-container">
-        <div class="col-12">
-            <span class="page-heading">Restriction</span>
-        </div>
-    </div>
-    <div class="row">
-        <div class="col-12">
-            <mat-card>
-                <p> Some Text</p>
-                <p> Some Text</p>
-                <p> Some Text</p>
-                <p> Some Text</p>
-                <p> Some Text</p>
-            </mat-card>
-        </div>
-    </div>
-    <br>
-    <br>
-    <br>
-    <div class="row">
-        <div class="col-12">
-            <button mat-flat-button color="primary" class="notice-button" (click)="submitRestriction()">Next - Create Account</button>
-        </div>
-    </div>
