--- evacuees/src/UI/embc-evacuee/src/app/restriction/restriction.component.scss
@@ -1,3 +1,0 @@
-.notice-button {
-    width: 100%;
-}
